requirements.txt
aiohttp>=3.7.4.post0
discord.py>=2.2.3